+++ app/scripts/i18n/index.ts
@@ -60,14 +60,14 @@ export function locObj(map?: LangString, lang?: string) {
  * Translate a given key in a translations list.
  * Very similar to `locObj(translations[key], lang)` but handles edge cases differently.
  * TODO Can we merge this with locObj?
- * @param {object} attribute An attribute config object. If it doesn't have a `translation` property, `key` is returned as is.
+ * @param {object} attribute An attribute config object. If omitted or it doesn't have a `translation` property, `key` is returned as is.
  * @param {string} key A translation key.
  * @param {string} [lang] The code of the language to translate to. Defaults to the global current language.
  * @returns {string} The translated string, undefined if no translation is found, or the value of `key` if `translations` is unusable.
  */
-export function locAttribute(attribute: Attribute, key: string, lang?: string): string {
+export function locAttribute(attribute: Attribute | undefined, key: string, lang?: string): string {
     lang = lang || getLang()
-    const translations = attribute.translation
+    const translations = attribute?.translation
     if (translations?.[key]) {
         const translation = translations[key]
         return isObject(translation) ? translation[lang] : translation